refactor: 시험 관련 mock 데이터 및 MSW handler 구조 단순화 (#66)
