--- src/mocks/data/exam/index.ts
@@ -1,3 +1,0 @@
-import { getExamListResponse } from "@/mocks/data/exam/exam-list";
-
-export { getExamListResponse };
